V2.2 update (#785)
* fix some code warnings

* pop stack warning for struct sockaddr_in since we are using it in the intended fashion but still generates warnings.

* fix ignore warning pragma

* fix a few warnings from boost 1.71 and visual studio 2019

* get rid of extraneous inclusion of boost

* update date and fmtlib

* Apply suggestions from code review

+++ src/helics/apps/Player.cpp
@@ -347,7 +347,7 @@ void Player::loadTextFile (const std::string &filename)
                 }
                 if (pIndex > 0)
                 {
-                    points[pIndex].pubName = points[pIndex - 1].pubName;
+                    points[pIndex].pubName = points[static_cast<size_t>(pIndex) - 1].pubName;
                 }
                 else
                 {
@@ -378,7 +378,7 @@ void Player::loadTextFile (const std::string &filename)
                 }
                 if ((blk[1].empty ()) && (pIndex > 0))
                 {
-                    points[pIndex].pubName = points[pIndex - 1].pubName;
+                    points[pIndex].pubName = points[static_cast<size_t>(pIndex) - 1].pubName;
                 }
                 else
                 {
@@ -409,7 +409,7 @@ void Player::loadTextFile (const std::string &filename)
                 }
                 if ((blk[1].empty ()) && (pIndex > 0))
                 {
-                    points[pIndex].pubName = points[pIndex - 1].pubName;
+                    points[pIndex].pubName = points[static_cast<size_t>(pIndex) - 1].pubName;
                 }
                 else
                 {
